# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple rate limiting and synchronization service for distributed systems",
        "title": "Bouncer API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.6"
    },
    "host": "localhost:5505",
    "basePath": "/",
    "paths": {
...
        "/barrier/{name}": {
            "delete": {
                "description": "Remove a barrier and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Barrier"
                ],
                "summary": "Delete a barrier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barrier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Barrier deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - barrier not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "x-order": 2
            }
        },
        "/barrier/{name}/wait": {
            "get": {
                "description": "Wait until N parties have arrived at the barrier",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Barrier"
                ],
                "summary": "Wait at barrier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barrier name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 2,
                        "description": "Number of parties to wait for",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": -1,
                        "description": "Maximum wait time",
                        "name": "maxwait",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Barrier completed successfully"
                    },
                    "408": {
                        "description": "Request timeout - maxWait exceeded",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "x-order": 1
            }
        },
        "/counter/{name}": {
            "delete": {
                "description": "Remove a counter and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Counter"
                ],
                "summary": "Delete a counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Counter deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - counter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counter/{name}/count": {
            "get": {
                "description": "Atomically adds amount to counter value",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Counter"
                ],
                "summary": "Increment or decrement counter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Amount to add (can be negative)",
                        "name": "amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New counter value",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/counter/{name}/reset": {
            "get": {
                "description": "Set counter to specified value",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Counter"
                ],
                "summary": "Reset counter value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Value to set",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Counter reset successful"
                    }
                }
            }
        },
        "/counter/{name}/value": {
            "get": {
                "description": "Returns current counter value",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Counter"
                ],
                "summary": "Get counter value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Counter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current counter value",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/{name}": {
            "delete": {
                "description": "Remove an event and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - event not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/{name}/send": {
            "get": {
                "description": "Trigger an event",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Send an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event sent successfully"
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - event handler not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict - event already sent",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/{name}/wait": {
            "get": {
                "description": "Wait for an event to be triggered",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Wait for an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": -1,
                        "description": "Maximum wait time",
                        "name": "maxwait",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event signal received",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - event handler not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}